Refactor GCP Scheduler to use shared components
