+++ jupyterlab_gcpscheduler/setup.cfg
@@ -1,5 +1,2 @@
 [bdist_wheel]
 universal=1
-
-[metadata]
-license_file = LICENSE
